Extract library to its own file
