+++ candle-metal-kernels/src/metal/mod.rs
@@ -1,4 +1,6 @@
 pub mod device;
+pub mod library;
 pub mod metal_utils;
 pub use device::*;
+pub use library::*;
 pub use metal_utils::*;
